chore: remove and move code
