+++ src/pipeline/tests/common.rs
@@ -13,13 +13,12 @@
 // limitations under the License.
 
 use greptime_proto::v1::{ColumnDataType, ColumnSchema, Rows, SemanticType};
-use pipeline::{
-    parse, serde_value_to_vrl_value, setup_pipeline, Content, Pipeline, PipelineContext,
-};
+use pipeline::{parse, setup_pipeline, Content, Pipeline, PipelineContext};
+use vrl::value::Value as VrlValue;
 
 /// test util function to parse and execute pipeline
 pub fn parse_and_exec(input_str: &str, pipeline_yaml: &str) -> Rows {
-    let input_value = serde_json::from_str::<serde_json::Value>(input_str).unwrap();
+    let input_value = serde_json::from_str::<VrlValue>(input_str).unwrap();
 
     let yaml_content = Content::Yaml(pipeline_yaml);
     let pipeline: Pipeline = parse(&yaml_content).expect("failed to parse pipeline");
@@ -34,21 +33,19 @@ pub fn parse_and_exec(input_str: &str, pipeline_yaml: &str) -> Rows {
     let mut rows = Vec::new();
 
     match input_value {
-        serde_json::Value::Array(array) => {
+        VrlValue::Array(array) => {
             for value in array {
-                let intermediate_status = serde_value_to_vrl_value(value).unwrap();
                 let row = pipeline
-                    .exec_mut(intermediate_status, &pipeline_ctx, &mut schema_info)
+                    .exec_mut(value, &pipeline_ctx, &mut schema_info)
                     .expect("failed to exec pipeline")
                     .into_transformed()
                     .expect("expect transformed result ");
                 rows.push(row.0);
             }
         }
-        serde_json::Value::Object(_) => {
-            let intermediate_status = serde_value_to_vrl_value(input_value).unwrap();
+        VrlValue::Object(_) => {
             let row = pipeline
-                .exec_mut(intermediate_status, &pipeline_ctx, &mut schema_info)
+                .exec_mut(input_value, &pipeline_ctx, &mut schema_info)
                 .expect("failed to exec pipeline")
                 .into_transformed()
                 .expect("expect transformed result ");
